Split resource world logger by world uuid, refactor
