+++ src/main/java/me/pugabyte/nexus/features/shops/MarketCommand.java
@@ -18,50 +18,25 @@ import me.pugabyte.nexus.framework.commands.models.events.CommandEvent;
 import me.pugabyte.nexus.framework.exceptions.postconfigured.InvalidInputException;
 import me.pugabyte.nexus.models.shop.ResourceMarketLogger;
 import me.pugabyte.nexus.models.shop.ResourceMarketLoggerService;
-import me.pugabyte.nexus.models.shop.Shop;
-import me.pugabyte.nexus.models.shop.Shop.BuyExchange;
-import me.pugabyte.nexus.models.shop.Shop.ExchangeType;
-import me.pugabyte.nexus.models.shop.Shop.Product;
-import me.pugabyte.nexus.models.shop.ShopService;
-import me.pugabyte.nexus.utils.ActionBarUtils;
 import me.pugabyte.nexus.utils.RandomUtils;
-import me.pugabyte.nexus.utils.Tasks;
 import me.pugabyte.nexus.utils.WorldEditUtils;
 import org.bukkit.Location;
-import org.bukkit.Material;
+import org.bukkit.World;
 import org.bukkit.block.Block;
-import org.bukkit.block.BlockFace;
-import org.bukkit.block.BlockState;
-import org.bukkit.entity.Player;
-import org.bukkit.entity.TNTPrimed;
-import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
-import org.bukkit.event.block.BlockBreakEvent;
-import org.bukkit.event.block.BlockDropItemEvent;
-import org.bukkit.event.block.BlockPistonExtendEvent;
-import org.bukkit.event.block.BlockPistonRetractEvent;
-import org.bukkit.event.block.BlockPlaceEvent;
-import org.bukkit.event.entity.EntityExplodeEvent;
-import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
 
-import java.math.BigDecimal;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
-
-import static eden.utils.StringUtils.prettyMoney;
-import static me.pugabyte.nexus.utils.ItemUtils.isNullOrAir;
 import static me.pugabyte.nexus.utils.WorldGroup.isResourceWorld;
 
 @NoArgsConstructor
 public class MarketCommand extends CustomCommand implements Listener {
+	private static final ResourceMarketLoggerService service = new ResourceMarketLoggerService();
+	private ResourceMarketLogger logger;
 
 	public MarketCommand(@NonNull CommandEvent event) {
 		super(event);
+		if (isPlayerCommandEvent())
+			logger = getLogger(world());
 	}
 
 	@Path
@@ -76,17 +51,15 @@ public class MarketCommand extends CustomCommand implements Listener {
 		send(PREFIX + "Market reloaded");
 	}
 
-	@NotNull
-	private ResourceMarketLoggerService getLoggerService() {
-		return new ResourceMarketLoggerService();
+	private ResourceMarketLogger getLogger(World world) {
+		if (!isResourceWorld(world))
+			throw new InvalidInputException("Not allowed outside of resource world");
+
+		return service.get(world.getUID());
 	}
 
-	private ResourceMarketLogger getLogger() {
-		return getLoggerService().get0();
-	}
-
-	private void save() {
-		getLoggerService().queueSave(Time.SECOND.get(), getLogger());
+	private void save(World world) {
+		service.queueSave(Time.SECOND.get(), getLogger(world));
 	}
 
 	@Async
@@ -100,10 +73,17 @@ public class MarketCommand extends CustomCommand implements Listener {
 			error("Max selection size is 1000000");
 
 		for (Block block : worldEditUtils.getBlocks(selection))
-			getLogger().add(block.getLocation());
+			logger.add(block.getLocation());
 
-		save();
-		send(PREFIX + "Added &e" + selection.getArea() + " &3locations to logger, new size: &e" + getLogger().size());
+		save(world());
+		send(PREFIX + "Added &e" + selection.getArea() + " &3locations to logger, new size: &e" + logger.size());
+	}
+
+	@Async
+	@Path("logger count")
+	@Permission("group.admin")
+	void logger_count() {
+		send(PREFIX + logger.size() + " coordinates logged");
 	}
 
 	@Async
@@ -114,7 +94,6 @@ public class MarketCommand extends CustomCommand implements Listener {
 		if (isResourceWorld(world()))
 			throw new InvalidInputException("You must be in a resource world");
 
-		final ResourceMarketLogger logger = getLogger();
 		for (int i = 0; i < amount; i++) {
 			while (true) {
 				final Location location = getRandomLocation();
@@ -126,8 +105,8 @@ public class MarketCommand extends CustomCommand implements Listener {
 			}
 		}
 
-		save();
-		send(PREFIX + "Added &e" + amount + " &3locations to logger, new size: &e" + getLogger().size());
+		save(world());
+		send(PREFIX + "Added &e" + amount + " &3locations to logger, new size: &e" + getLogger(world()).size());
 	}
 
 	@NotNull
@@ -136,146 +115,6 @@ public class MarketCommand extends CustomCommand implements Listener {
 		final int y = RandomUtils.randomInt(-64, 319);
 		final int z = RandomUtils.randomInt(-ResourceWorld.RADIUS, ResourceWorld.RADIUS);
 		return new Location(world(), x, y, z);
-	}
-
-	@Async
-	@Path("logger count")
-	@Permission("group.admin")
-	void logger_count() {
-		send(PREFIX + getLogger().size() + " coordinates logged");
-	}
-
-	@EventHandler
-	public void onBlockPlace(BlockPlaceEvent event) {
-		if (!isResourceWorld(event.getBlock()))
-			return;
-
-		getLogger().add(event.getBlock().getLocation());
-		save();
-	}
-
-	@EventHandler
-	public void onBlockBreak(BlockBreakEvent event) {
-		if (!isResourceWorld(event.getBlock()))
-			return;
-
-		getLogger().remove(event.getBlock().getLocation());
-		save();
-	}
-
-	@EventHandler
-	public void onBlockPistonExtend(BlockPistonExtendEvent event) {
-		handlePiston(event.getBlock(), event.getBlocks(), event.getDirection());
-	}
-
-	@EventHandler
-	public void onBlockPistonExtend(BlockPistonRetractEvent event) {
-		handlePiston(event.getBlock(), event.getBlocks(), event.getDirection());
-	}
-
-	private void handlePiston(Block eventBlock, List<Block> blocks, BlockFace direction) {
-		if (!isResourceWorld(eventBlock))
-			return;
-
-		getLogger().add(eventBlock.getLocation());
-		for (Block block : blocks) {
-			getLogger().add(block.getLocation());
-			getLogger().add(block.getRelative(direction).getLocation());
-		}
-
-		save();
-	}
-
-	@EventHandler
-	public void onBlockDropItem(BlockDropItemEvent event) {
-		if (!isResourceWorld(event.getBlock()))
-			return;
-
-		event.getItems().removeIf(item ->
-			trySell(event.getPlayer(), event.getBlockState(), item.getItemStack()));
-	}
-
-	@EventHandler
-	public void onEntityExplode(EntityExplodeEvent event) {
-		if (!isResourceWorld(event.getLocation()))
-			return;
-
-		if (!(event.getEntity() instanceof TNTPrimed tnt))
-			return;
-
-		if (!(tnt.getSource() instanceof Player player))
-			return;
-
-		final Iterator<Block> iterator = event.blockList().iterator();
-		while (iterator.hasNext()) {
-			final Block next = iterator.next();
-			for (ItemStack drop : next.getDrops())
-				if (trySell(player, next.getState(), drop)) {
-					next.setType(Material.AIR);
-					iterator.remove();
-				}
-		}
-	}
-
-	private boolean trySell(Player player, BlockState block, ItemStack drop) {
-		if (isNullOrAir(block.getType()) || isNullOrAir(drop))
-			return false;
-
-		final Shop shopper = new ShopService().get(player);
-		final boolean disabled = shopper.getDisabledResourceMarketItems().contains(drop.getType());
-		if (disabled)
-			return false;
-
-		if (getLogger().contains(block.getLocation()))
-			return false;
-
-		return trySell(player, drop);
-	}
-
-	private static final Map<UUID, BigDecimal> earned = new HashMap<>();
-	private static final Map<UUID, Integer> taskIds = new HashMap<>();
-
-	private void actionBar(Player player) {
-		final UUID uuid = player.getUniqueId();
-		final BigDecimal number = earned.get(uuid);
-
-		if (number.signum() != 0) {
-			Tasks.cancel(taskIds.getOrDefault(uuid, -1));
-			ActionBarUtils.sendActionBar(player, "&a+" + prettyMoney(number));
-			final int taskId = Tasks.wait(Time.SECOND.x(3.5), () -> earned.put(uuid, new BigDecimal(0)));
-			taskIds.put(uuid, taskId);
-		}
-	}
-
-	private boolean trySell(Player player, ItemStack item) {
-		final Optional<Product> product = getMatchingProduct(item);
-		if (product.isEmpty())
-			return false;
-
-		if (!(product.get().getExchange() instanceof BuyExchange exchange))
-			throw new InvalidInputException("Cannot process resource market exchange: " + camelCase(product.get().getExchangeType()));
-
-		process(player, item, exchange);
-		actionBar(player);
-
-		return true;
-	}
-
-	private void process(Player player, ItemStack item, BuyExchange exchange) {
-		BigDecimal thing = earned.getOrDefault(player.getUniqueId(), new BigDecimal(0));
-		earned.put(player.getUniqueId(), thing.add(exchange.processResourceMarket(player, item)));
-	}
-
-	private Optional<Product> getMatchingProduct(ItemStack item) {
-		return getResourceWorldProducts().stream()
-			.filter(product -> product.getItem().isSimilar(item))
-			.findFirst();
-	}
-
-	private List<Product> getResourceWorldProducts() {
-		return new ShopService().getMarket().getProducts().stream()
-			.filter(product -> product.isResourceWorld() && product.getExchangeType() == ExchangeType.BUY)
-			.toList();
 	}
 
 }
